{"schema":"decodex/commit/1","summary":"Modularize window renderer surface lifecycle","authority":"manual"}
